# src/albums/entities/album.entity.ts
export class Album {
  id: string; // uuid v4
  name: string;
  year: number;
  artistId: string | null; // refers to Artist

  constructor(partial: Partial<Album>) {
    Object.assign(this, partial);
    this.id = crypto.randomUUID();
  }
}