# src/artists/entities/artist.entity.ts
export class Artist {
  id: string; // uuid v4
  name: string;
  grammy: boolean;

  constructor(partial: Partial<Artist>) {
    Object.assign(this, partial);
    this.id = crypto.randomUUID();
  }
}